@@ -3,32 +3,57 @@ name = "automatic-diary"
 version = "0.4.0"
 description = "Create one timeline from various digital sources discribing your life."
 authors = ["Jakub Valenta <jakub@jakubvalenta.cz>"]
-license = "GNU General Public License v3 or later (GPLv3+)"
+license = "GPLv3+"
+readme = "README.md"
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: End Users/Desktop",
+    "Topic :: Utilities",
+    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12"
+]
 
 [tool.poetry.dependencies]
-python = "^3.7"
-beautifulsoup4 = "^4.10.0"
-caldav = "^0.8.0"
-dateparser = "^1.0.0"
+Jinja2 = "^3.1.4"
+beautifulsoup4 = "^4.12.3"
+caldav = "^1.3.9"
+dateparser = "^1.2.0"
 ics = {git = "https://github.com/ics-py/ics-py.git", branch = "main"}
-Jinja2 = "^3.0.1"
-more-itertools = "^8.10.0"
-pystache = "^0.5.4"
-python-dateutil = "^2.8.2"
-requests = "^2.26.0"
-tzlocal = "^3.0"
-orgparse = "^0.2.4"
+lxml = "^5.3.0"
+more-itertools = "^10.5.0"
+orgparse = "^0.4.20231004"
+pystache = "^0.6.5"
+python = "^3.9"
+python-dateutil = "^2.9.0.post0"
+requests = "^2.32.3"
 "trakt.py" = "^4.4.0"
+tzlocal = "^5.2"
 
-[tool.poetry.dev-dependencies]
-ddt = "^1.4.2"
-flake8 = "^3.9.2"
-isort = "^5.9.3"
-mypy = "^0.910"
-types-requests = "^2.25.6"
-types-python-dateutil = "^2.8.0"
-types-dateparser = "^1.0.2"
+[tool.poetry.scripts]
+fit2gpx = "fit2gpx:cli"
+automatic-diary = "automatic_diary.cli:main"
+automatic-diary-visualize = "automatic_diary.visualize:main"
+
+[tool.poetry.group.dev.dependencies]
+ddt = "^1.7.2"
+isort = "^5.13.2"
+mypy = "^1.11.2"
+types-requests = "^2.32.0.20240914"
+types-python-dateutil = "^2.9.0.20240906"
+types-dateparser = "^1.2.0.20240420"
+flake8-pyproject = "^1.2.3"
+
+[tool.flake8]
+max-line-length = 100
+
+[tool.isort]
+profile = "django"
+line_length = 100
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
+requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
