@@ -2,7 +2,7 @@ import datetime
 import logging
 import re
 from pathlib import Path
-from typing import IO, Iterator, List, Optional
+from typing import IO, Iterator, Optional
 
 from more_itertools import peekable
 
@@ -11,12 +11,12 @@ from automatic_diary.model import Item
 logger = logging.getLogger(__name__)
 provider = Path(__file__).parent.name
 
-regex_heading = re.compile(r'^\* (?P<todo>TODO )?[<\[](?P<date>.+)[>\]]$')
+regex_heading = re.compile(r"^\* (?P<todo>TODO )?[<\[](?P<date>.+)[>\]]$")
 
 
 def parse_orgmode(f: IO, subprovider: str) -> Iterator[Item]:
     current_datetime: Optional[datetime.datetime] = None
-    current_paragraph: List[str] = []
+    current_paragraph: list[str] = []
     lines = peekable(f)
     for line in lines:
         line_clean = line.strip()
@@ -24,12 +24,10 @@ def parse_orgmode(f: IO, subprovider: str) -> Iterator[Item]:
             m = regex_heading.match(line_clean)
             # Title line
             if m:
-                if m.group('todo'):
+                if m.group("todo"):
                     current_datetime = None
                 else:
-                    current_datetime = datetime.datetime.strptime(
-                        m.group('date'), '%Y-%m-%d %a'
-                    )
+                    current_datetime = datetime.datetime.strptime(m.group("date"), "%Y-%m-%d %a")
             # Paragraph line but not before first heading
             elif current_datetime:
                 current_paragraph.append(line_clean)
@@ -38,7 +36,7 @@ def parse_orgmode(f: IO, subprovider: str) -> Iterator[Item]:
             if current_datetime and current_paragraph:
                 yield Item.normalized(
                     datetime_=current_datetime,
-                    text='\n'.join(current_paragraph),
+                    text="\n".join(current_paragraph),
                     provider=provider,
                     subprovider=subprovider,
                     all_day=True,
@@ -47,8 +45,8 @@ def parse_orgmode(f: IO, subprovider: str) -> Iterator[Item]:
 
 
 def main(config: dict, *args, **kwargs) -> Iterator[Item]:
-    path = Path(config['path'])
+    path = Path(config["path"])
     subprovider = path.name
-    logger.info('Reading Org-mode file %s', path)
+    logger.info("Reading Org-mode file %s", path)
     with path.open() as f:
         yield from parse_orgmode(f, subprovider)
