@@ -1,8 +1,8 @@
 import subprocess
-from typing import List, Optional
+from typing import Optional
 
 
-def run_shell_cmd(cmd: List[str], **kwargs) -> str:
+def run_shell_cmd(cmd: list[str], **kwargs) -> str:
     completed_process = subprocess.run(
         cmd,
         stdout=subprocess.PIPE,
@@ -14,15 +14,15 @@ def run_shell_cmd(cmd: List[str], **kwargs) -> str:
 
 
 def search_secret(key: str, val: str, label: str) -> Optional[str]:
-    out = run_shell_cmd(['secret-tool', 'search', key, val])
+    out = run_shell_cmd(["secret-tool", "search", key, val])
     lines = out.splitlines()
     for i, line in enumerate(lines):
-        if line == f'label = {label}':
+        if line == f"label = {label}":
             if len(lines) == i + 1:
-                raise Exception('Invalid secret-tool output')
+                raise Exception("Invalid secret-tool output")
             secret_line = lines[i + 1]
-            if not secret_line.startswith('secret = '):
-                raise Exception('Invalid secret-tool output')
-            start = len('secret = ')
+            if not secret_line.startswith("secret = "):
+                raise Exception("Invalid secret-tool output")
+            start = len("secret = ")
             return secret_line[start:]
     return None
