@@ -8,7 +8,7 @@ import sys
 from collections import defaultdict
 from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Dict, Iterable, Iterator, List
+from typing import Iterable, Iterator
 
 from jinja2 import Environment, PackageLoader, select_autoescape
 
@@ -24,10 +24,10 @@ class Day:
     date: datetime.date
     today: bool
     even: bool
-    items: List[Item] = field(default_factory=list)
+    items: list[Item] = field(default_factory=list)
 
     @classmethod
-    def from_date(cls, date: datetime.date, *args, **kwargs) -> 'Day':
+    def from_date(cls, date: datetime.date, *args, **kwargs) -> "Day":
         today_ = date == today
         even = bool(date.month % 2)
         return cls(date, today_, even, *args, **kwargs)
@@ -37,12 +37,10 @@ class Week(list):
     pass
 
 
-def _create_days_around(
-    date: datetime.date, start: int, stop: int
-) -> Iterator[Day]:
+def _create_days_around(date: datetime.date, start: int, stop: int) -> Iterator[Day]:
     for i in range(start, stop):
         empty_date = date + datetime.timedelta(days=i)
-        logger.info('Empty day %s', empty_date)
+        logger.info("Empty day %s", empty_date)
         yield Day.from_date(empty_date)
 
 
@@ -63,10 +61,8 @@ def _group_items_in_days(items: Iterable[Item]) -> Iterator[Day]:
             current_date = date
         elif date != current_date:
             yield Day.from_date(current_date, current_items)
-            yield from _create_days_around(
-                current_date, start=1, stop=(date - current_date).days
-            )
-            logger.info('New day %s', date)
+            yield from _create_days_around(current_date, start=1, stop=(date - current_date).days)
+            logger.info("New day %s", date)
             current_date = date
             current_items = []
         current_items.append(item)
@@ -78,20 +74,19 @@ def _calc_perc(part: float, whole: float) -> int:
     return round(min(part / whole, 1) * 100)
 
 
-def _calc_stats(weeks: Iterable[Week]) -> List[Dict[str, int]]:
-    provider_counts_by_week: List[Dict[str, int]] = []
-    provider_counts: Dict[str, List[int]] = defaultdict(list)
+def _calc_stats(weeks: Iterable[Week]) -> list[dict[str, int]]:
+    provider_counts_by_week: list[dict[str, int]] = []
+    provider_counts: dict[str, list[int]] = defaultdict(list)
     for week in weeks:
-        week_provider_counts: Dict[str, int] = defaultdict(int)
+        week_provider_counts: dict[str, int] = defaultdict(int)
         for day in week:
             for item in day.items:
                 week_provider_counts[item.provider] += 1
         for provider, counts in week_provider_counts.items():
             provider_counts[provider].append(counts)
         provider_counts_by_week.append(week_provider_counts)
-    provider_means: Dict[str, float] = {
-        provider: statistics.mean(counts)
-        for provider, counts in provider_counts.items()
+    provider_means: dict[str, float] = {
+        provider: statistics.mean(counts) for provider, counts in provider_counts.items()
     }
     stats = [
         {
@@ -120,17 +115,15 @@ def _matches_regex(s: str, regex: str) -> bool:
     return re.search(regex, s) is not None
 
 
-def _render_template(
-    package: List[str], output_html_path: str, highlight: str, **context
-):
+def _render_template(package: list[str], output_html_path: str, highlight: str, **context):
     environment = Environment(
-        autoescape=select_autoescape(['html']),
+        autoescape=select_autoescape(["html"]),
         loader=PackageLoader(*package[:-1]),
     )
-    environment.tests['highlighted'] = lambda s: _matches_regex(s, highlight)
+    environment.tests["highlighted"] = lambda s: _matches_regex(s, highlight)
     template = environment.get_template(package[-1])
     stream = template.stream(**context)
-    with Path(output_html_path).open('w') as f:
+    with Path(output_html_path).open("w") as f:
         f.writelines(stream)
 
 
@@ -140,7 +133,7 @@ def _visualize(csv_path: str, output_html_path: str, highlight: str):
     weeks = list(_group_days_in_weeks(days))
     stats = _calc_stats(weeks)
     _render_template(
-        ['automatic_diary', 'templates', 'template.html'],
+        ["automatic_diary", "templates", "template.html"],
         output_html_path,
         highlight,
         weeks=weeks,
@@ -149,24 +142,16 @@ def _visualize(csv_path: str, output_html_path: str, highlight: str):
 
 
 def main():
-    parser = argparse.ArgumentParser(
-        description='Visualize Automatic Diary CSV'
-    )
-    parser.add_argument('csv_path', help='Input CSV file path')
-    parser.add_argument('output_html_path', help='Output HTML file path')
-    parser.add_argument(
-        '-i', '--highlight', help='Highlight items matching regex'
-    )
-    parser.add_argument(
-        '-v', '--verbose', action='store_true', help='Enable debugging output'
-    )
+    parser = argparse.ArgumentParser(description="Visualize Automatic Diary CSV")
+    parser.add_argument("csv_path", help="Input CSV file path")
+    parser.add_argument("output_html_path", help="Output HTML file path")
+    parser.add_argument("-i", "--highlight", help="Highlight items matching regex")
+    parser.add_argument("-v", "--verbose", action="store_true", help="Enable debugging output")
     args = parser.parse_args()
     if args.verbose:
-        logging.basicConfig(
-            stream=sys.stdout, level=logging.INFO, format='%(message)s'
-        )
+        logging.basicConfig(stream=sys.stdout, level=logging.INFO, format="%(message)s")
     _visualize(args.csv_path, args.output_html_path, args.highlight)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()
