@@ -1,4 +1,4 @@
 from automatic_diary.cli import main
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()
